Moved hotkeys models to Common. Added partial support for hotkeys settings.
