+++ src/Osu.Music.UI/Views/SettingsView.xaml
@@ -7,12 +7,12 @@
              xmlns:prism="http://prismlibrary.com/"
              xmlns:local="clr-namespace:Osu.Music.UI.Views"
              xmlns:colorpicker="clr-namespace:ColorPicker;assembly=ColorPicker"
-             xmlns:hk="clr-namespace:Osu.Music.Services.Hotkeys;assembly=Osu.Music.Services"
+             xmlns:models="clr-namespace:Osu.Music.Common.Models;assembly=Osu.Music.Common"
              xmlns:uc="clr-namespace:Osu.Music.UI.UserControls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <UserControl.Resources>
-        <DataTemplate x:Key="HotkeyTemplate" DataType="{x:Type hk:Hotkey}">
+        <DataTemplate x:Key="HotkeyTemplate" DataType="{x:Type models:Hotkey}">
             <uc:HotkeyListItem Margin="0 5 0 5" SelectedHotkey="{Binding DataContext.SelectedHotkey, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:SettingsView}}}"/>
         </DataTemplate>
     </UserControl.Resources>
